# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Coffy Shop API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MenuCategory {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
}

export interface MenuItem {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  price: number;
  categoryId: number;
  /** @nullable */
  categoryName?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  available: boolean;
  featured: boolean;
}

export type StoreStatsItemsPerCategoryItem = {
  categoryName: string;
  count: number;
};

export interface StoreStats {
  totalItems: number;
  totalCategories: number;
  featuredCount: number;
  itemsPerCategory?: StoreStatsItemsPerCategoryItem[];
}

export interface OrderLineItem {
  menuItemId: number;
  name: string;
  quantity: number;
  price: number;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  pending: 'pending',
  paid: 'paid',
  failed: 'failed',
  cancelled: 'cancelled',
} as const;

export interface Order {
  id: number;
  customerName: string;
  customerEmail: string;
  items: OrderLineItem[];
  total: number;
  status: OrderStatus;
  /** @nullable */
  stripePaymentIntentId?: string | null;
  createdAt: string;
}

export interface OrderInput {
  /** @minLength 1 */
  customerName: string;
  /** @minLength 1 */
  customerEmail: string;
  items: OrderLineItem[];
}

export interface PaymentIntentInput {
  orderId: number;
}

export interface PaymentIntentResult {
  clientSecret: string;
  paymentIntentId: string;
  amount: number;
}

export interface PaymentConfirmInput {
  orderId: number;
  paymentIntentId: string;
}

export type ListMenuItemsParams = {
categoryId?: number;
};