# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Coffy Shop API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: menu
    description: Menu items and categories
  - name: orders
    description: Order management
  - name: payment
    description: Stripe payment operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /menu-categories:
    get:
      operationId: listMenuCategories
      tags: [menu]
      summary: List all menu categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MenuCategory"

  /menu-items:
    get:
      operationId: listMenuItems
      tags: [menu]
      summary: List all menu items, optionally filtered by category
      parameters:
        - name: categoryId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of menu items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MenuItem"

  /menu-items/featured:
    get:
      operationId: listFeaturedMenuItems
      tags: [menu]
      summary: Get featured menu items for the homepage
      responses:
        "200":
          description: Featured menu items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MenuItem"

  /menu-items/{id}:
    get:
      operationId: getMenuItem
      tags: [menu]
      summary: Get a single menu item by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Menu item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /store-stats:
    get:
      operationId: getStoreStats
      tags: [menu]
      summary: Get store stats (item counts, category breakdown)
      responses:
        "200":
          description: Store statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoreStats"

  /orders:
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get an order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /payment/create-intent:
    post:
      operationId: createPaymentIntent
      tags: [payment]
      summary: Create a Stripe payment intent for an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentIntentInput"
      responses:
        "200":
          description: Payment intent created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentIntentResult"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /payment/confirm:
    post:
      operationId: confirmPayment
      tags: [payment]
      summary: Confirm payment and update order status
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentConfirmInput"
      responses:
        "200":
          description: Payment confirmed, order updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    MenuCategory:
      type: object
      required: [id, name, slug]
      properties:
        id:
          type: number
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]

    MenuItem:
      type: object
      required: [id, name, price, categoryId, available, featured]
      properties:
        id:
          type: number
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        categoryId:
          type: number
        categoryName:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        available:
          type: boolean
        featured:
          type: boolean

    StoreStats:
      type: object
      required: [totalItems, totalCategories, featuredCount]
      properties:
        totalItems:
          type: number
        totalCategories:
          type: number
        featuredCount:
          type: number
        itemsPerCategory:
          type: array
          items:
            type: object
            required: [categoryName, count]
            properties:
              categoryName:
                type: string
              count:
                type: number

    OrderLineItem:
      type: object
      required: [menuItemId, quantity, price, name]
      properties:
        menuItemId:
          type: number
        name:
          type: string
        quantity:
          type: number
        price:
          type: number

    Order:
      type: object
      required: [id, customerName, customerEmail, items, total, status, createdAt]
      properties:
        id:
          type: number
        customerName:
          type: string
        customerEmail:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderLineItem"
        total:
          type: number
        status:
          type: string
          enum: [pending, paid, failed, cancelled]
        stripePaymentIntentId:
          type: ["string", "null"]
        createdAt:
          type: string

    OrderInput:
      type: object
      required: [customerName, customerEmail, items]
      properties:
        customerName:
          type: string
          minLength: 1
        customerEmail:
          type: string
          minLength: 1
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderLineItem"

    PaymentIntentInput:
      type: object
      required: [orderId]
      properties:
        orderId:
          type: number

    PaymentIntentResult:
      type: object
      required: [clientSecret, paymentIntentId, amount]
      properties:
        clientSecret:
          type: string
        paymentIntentId:
          type: string
        amount:
          type: number

    PaymentConfirmInput:
      type: object
      required: [orderId, paymentIntentId]
      properties:
        orderId:
          type: number
        paymentIntentId:
          type: string