# lib/api-zod/src/generated/types/menuCategory.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Coffy Shop API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MenuCategory {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
}