# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Coffy Shop API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderLineItem } from './orderLineItem';
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: number;
  customerName: string;
  customerEmail: string;
  items: OrderLineItem[];
  total: number;
  status: OrderStatus;
  /** @nullable */
  stripePaymentIntentId?: string | null;
  createdAt: string;
}