# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Coffy Shop API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderLineItem } from './orderLineItem';

export interface OrderInput {
  /** @minLength 1 */
  customerName: string;
  /** @minLength 1 */
  customerEmail: string;
  items: OrderLineItem[];
}