# lib/api-zod/src/generated/types/orderLineItem.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Coffy Shop API specification
 * OpenAPI spec version: 0.1.0
 */

export interface OrderLineItem {
  menuItemId: number;
  name: string;
  quantity: number;
  price: number;
}