# lib/api-zod/src/generated/types/storeStats.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Coffy Shop API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StoreStatsItemsPerCategoryItem } from './storeStatsItemsPerCategoryItem';

export interface StoreStats {
  totalItems: number;
  totalCategories: number;
  featuredCount: number;
  itemsPerCategory?: StoreStatsItemsPerCategoryItem[];
}